ENH: Add to_latex() method to Series (#16180)
This changeset adds _repr_latex_ to the Series class and moves
the to_latex() method from the DataFrame class to the NDFrame
class.
